+++ src/modules/dateTimePicker/components/index.ts
@@ -1,0 +1,2 @@
+export * from './dateTimePicker/dateTimePicker.component';
+export * from './dayPicker/dayPicker.component';