+++ src/modules/dateTimePicker/directives/index.ts
@@ -1,0 +1,2 @@
+export * from './dateTimePicker/dateTimePicker.directive';
+export * from './dateTimePicker/dateTimePicker.interface';