+++ src/modules/index.ts
@@ -1,2 +1,3 @@
 export * from './datePipes.module';
 export * from './dateTime';
+export * from './dateTimePicker';